# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.HealthResponse:
    properties:
      message:
        example: Service is running
        type: string
      status:
        example: ok
        type: string
    type: object
  models.AudioDebate:
    properties:
      cloudinary_audio_url:
        type: string
      created_at:
        type: string
      document_id:
        type: integer
      id:
        type: integer
      transcript_json:
        type: string
      updated_at:
        type: string
    type: object
  models.AuditReport:
    properties:
      created_at:
        type: string
      document_id:
        type: integer
      fortification_plan:
        type: string
      heatmap_data:
        type: string
      id:
        type: integer
      logical_fallacies:
        type: string
      resilience_score:
        type: integer
      updated_at:
        type: string
      vulnerabilities:
        type: string
    type: object
  models.Document:
    properties:
      audio_debate:
        $ref: '#/definitions/models.AudioDebate'
      audit_report:
        $ref: '#/definitions/models.AuditReport'
      cloudinary_url:
        type: string
      created_at:
        type: string
      file_name:
        type: string
      id:
        type: integer
      raw_text:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      message:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.Payment:
    properties:
      amount_paid:
        type: number
      created_at:
        type: string
      id:
        type: integer
      status:
        type: string
      stripe_session_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      operator_name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - operator_name
    - password
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        type: object
    type: object
  models.UpdateDocumentRequest:
    properties:
      file_name:
        type: string
      status:
        type: string
    type: object
  models.UpdateUserRequest:
    properties:
      email:
        type: string
      status:
        enum:
        - active
        - suspended
        type: string
      subscription_tier:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      documents:
        items:
          $ref: '#/definitions/models.Document'
        type: array
      email:
        type: string
      id:
        type: integer
      operator_name:
        type: string
      payments:
        items:
          $ref: '#/definitions/models.Payment'
        type: array
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      status:
        type: string
      stripe_customer_id:
        type: string
      subscription_tier:
        type: string
      updated_at:
        type: string
      user_usage:
        $ref: '#/definitions/models.UserUsage'
    type: object
  models.UserUsage:
    properties:
      audit_limit:
        type: integer
      audits_performed:
        type: integer
      created_at:
        type: string
      last_reset_at:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:3000
info:
  contact:
    email: support@ares-ai.com
    name: API Support
  description: Stress-Test Engine Backend - AI-powered document analysis and resilience
    scoring
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ares AI API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      description: Logout the current user (clears session)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout user
      tags:
      - Auth
  /auth/me:
    get:
      description: Get the currently authenticated user's details with role information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current user
      tags:
      - Auth
  /auth/refresh:
    post:
      description: Generate a new JWT access token using a valid refresh token from
        cookies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account. First user gets Admin role, subsequent
        users get User role.
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
  /documents:
    get:
      description: Retrieve all documents in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Document'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all documents
      tags:
      - Documents
    post:
      consumes:
      - multipart/form-data
      description: Create a new document and upload file to Cloudinary
      parameters:
      - description: Document file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Document'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create document with file upload
      tags:
      - Documents
  /documents/{id}:
    delete:
      description: Remove a document (users can only delete their own, admins can
        delete any)
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete document
      tags:
      - Documents
    get:
      description: Retrieve a single document's details (users can only see their
        own, admins can see any)
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Document'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get document by ID
      tags:
      - Documents
    put:
      consumes:
      - application/json
      description: Update document information (users can only update their own, admins
        can update any)
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Document'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update document
      tags:
      - Documents
  /documents/user/{userId}:
    get:
      description: Retrieve documents for the current user (users see only their own,
        admins can request any user)
      parameters:
      - description: User ID (optional for admins)
        in: path
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Document'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user documents
      tags:
      - Documents
  /health:
    get:
      consumes:
      - application/json
      description: Check if the API service is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HealthResponse'
      summary: Health check endpoint
      tags:
      - Health
  /users:
    get:
      description: Retrieve a list of all users with their role information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all users
      tags:
      - Users
  /users/{id}:
    delete:
      description: Remove a user and all related data (users can only delete their
        own, admins can delete any) (CASCADE)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user
      tags:
      - Users
    get:
      description: Retrieve a single user's details by their ID with role information
        (users can only see their own, admins can see any)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user's email, subscription tier, or status (users can only
        update their own, admins can update any)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user
      tags:
      - Users
schemes:
- http
- https
swagger: "2.0"